Move terminalWidgetManager to browser/
